workers/requirements.txt
# Python Dependencies para Pipeline de Clonagem de Voz Profissional
# Instalar com: python -m pip install -r requirements.txt
# 
# NOTA: Python 3.14 não é suportado por algumas bibliotecas
# Use Python 3.11 ou 3.12 para melhor compatibilidade

# Processamento de Áudio
librosa>=0.10.0,<0.11.0
soundfile>=0.12.0
noisereduce>=3.0.0
pydub>=0.25.1

# Extração de Embeddings de Voz
# Resemblyzer pode ter problemas com Python 3.14
# Se falhar, tente instalar versão mais recente ou use alternativa
resemblyzer>=0.1.1

# Computação Numérica (essencial)
numpy>=1.24.0,<2.0.0
scipy>=1.11.0

# SpeechBrain (Opcional - para ECAPA-TDNN, requer GPU)
# speechbrain>=0.5.15
# torch>=2.0.0
# torchaudio>=2.0.0

# Coqui TTS - Text-to-Speech
# NOTA: Para Python 3.12, use o fork coqui-tts (suporta Python 3.12)
# Para Python 3.11 ou anterior, use TTS>=0.22.0
coqui-tts>=0.1.0  # Fork que suporta Python 3.12
torch>=2.0.0
torchaudio>=2.0.0

# Utilitários
requests>=2.31.0
python-dotenv>=1.0.0
